requirements.txt
tensorflow
numpy
matplotlib
pillow
requests
tk
